# Quantity input speed