--- a/astropy/units/decorators.py
+++ b/astropy/units/decorators.py
@@ -59,9 +59,9 @@ def _validate_arg_value(
     # If dimensionless is an allowed unit and the argument is unit-less,
     #   allow numbers or numpy arrays with numeric dtypes
     if (
-        dimensionless_unscaled in allowed_units
-        and not strict_dimensionless
+        not strict_dimensionless
         and not hasattr(arg, "unit")
+        and dimensionless_unscaled in allowed_units
     ):
         if isinstance(arg, Number):
             return