@svank 2024-07-18
Good idea! How does this look?