@pllim 2024-07-18
Might be a fluke, not sure.

```
| Change   | Before [16c7f4d7]    | After [29219ec4]    |   Ratio | Benchmark (Parameter)          |
|----------|----------------------|---------------------|---------|--------------------------------|
| +        | 45.1±0.4μs           | 102±8μs             |    2.26 | units.time_quantity_times_unit |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```