@mhvk 2024-07-19
Our new `perf` changelog entry is specifically for cases like these with a clear improvement for user-facing API. So, yes, lets add a `docs/changes/units/16742.perf.rst`!